+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - DF-920 Allowing SMTP service without authentication.
 - DF-447 Adding support for Azure Active Directory.
+- DF-924 Adding support for event-driven logging service.
 
 ### Changed
 - Handling errors and showing original content when content in response cannot be resolved to Accept type.